+++ src/util.h
@@ -19,4 +19,4 @@
  *
  */
 
-int util_config_is_valid(struct l_queue *config_list);
+int8_t util_config_is_valid(struct l_queue *config_list);
